Avoid `defineNuxtPlugin()` in a tested file
